+++ .gitignore
@@ -13,3 +13,4 @@
 /phpunit.xml
 .phpunit.result.cache
 ###< phpunit/phpunit ###
+/config/environments/.env.tests
