+++ bin/console
@@ -1,7 +1,7 @@
 #!/usr/bin/env php
 <?php
 
-use Cmas\MyApp\Kernel;
+use Cmas\MyApp\Apps\Kernel;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 
 if (!is_file(dirname(__DIR__).'/vendor/autoload_runtime.php')) {
