+++ config/environments/.env.example
@@ -1,0 +1,2 @@
+APP_ENV=dev
+APP_SECRET=5cd657db553232a29227e75da8472287
