+++ config/services.yaml
@@ -17,7 +17,7 @@ services:
         resource: '../src/'
         exclude:
 
-    Cmas\Apps\:
+    Cmas\MyApp\Apps\:
         resource: '../apps/'
 
     # add more service definitions when explicit configuration is needed
