+++ public/index.php
@@ -1,9 +1,9 @@
 <?php
 
-use Cmas\MyApp\Kernel;
+use Cmas\MyApp\Apps\Kernel;
 
 require_once dirname(__DIR__).'/vendor/autoload_runtime.php';
 
-return function (array $context) {
+return static function (array $context) {
     return new Kernel($context['APP_ENV'], (bool) $context['APP_DEBUG']);
 };
